pyproject.toml
[tool.poetry]
name = "pytest-dbt-postgres"
version = "0.1.0"
description = "Pytest tooling to unittest DBT & Postgres models"
authors = ["albert.franzi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.3"
pytest-env = "^1.1.3"
pytest-postgresql = "^5.0.0"
dbt-core = "^1.7.4"
dbt-postgres = "^1.7.4"
numpy = "^1.26.2"
pandas = "^2.1.4"
absl-py = "^2.0.0"
pydantic = ">=1.10,<2.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
pytest-cov = "^4.1.0"
tox = "^4.11.4"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-p no:warnings"
testpaths = ["tests"]